to widget that coordinator layout optimized
