+++ app/src/main/java/com/dvsnier/wrapper/TransferStationWrapper.java
@@ -15,9 +15,9 @@ import com.dvsnier.test.view.affinity.AffinityActivity;
 import com.dvsnier.test.widget.theme.TestThemeActivity;
 import com.dvsnier.testAIDL.TestAIDLActivity;
 import com.dvsnier.test.widget.animator.TestAnimator;
-import com.dvsnier.testCoordinatorLayout.TestCoordinatorLayout2Activity;
-import com.dvsnier.testCoordinatorLayout.TestCoordinatorLayout3Activity;
-import com.dvsnier.testCoordinatorLayout.TestCoordinatorLayoutActivity;
+import com.dvsnier.test.widget.coordinator.TestCoordinatorLayout2Activity;
+import com.dvsnier.test.widget.coordinator.TestCoordinatorLayout3Activity;
+import com.dvsnier.test.widget.coordinator.TestCoordinatorLayoutActivity;
 import com.dvsnier.testEventBus.TestEventBusActivity;
 import com.dvsnier.testGreenDao.TestGreenDaoActivity;
 import com.dvsnier.testImage.TestImageActivity;
