+++ research/widget/src/main/java/com/dvsnier/test/widget/coordinator/TestCoordinatorLayout3Activity.java
@@ -1,9 +1,9 @@
-package com.dvsnier.testCoordinatorLayout;
+package com.dvsnier.test.widget.coordinator;
 
 import android.os.Bundle;
 
-import com.dvsnier.R;
 import com.dvsnier.base.flavor.activity.BaseActivity;
+import com.dvsnier.test.widget.R;
 
 import butterknife.ButterKnife;
