+++ research/widget/src/main/res/drawable-hdpi/tool_bar.png
