+++ research/widget/src/main/res/drawable-xhdpi/tool_bar.png
