+++ research/widget/src/main/res/drawable-xxhdpi/tool_bar.png
