+++ research/widget/src/main/res/layout/activity_test_coordinator_layout.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context="com.dvsnier.test.widget.coordinator.TestCoordinatorLayoutActivity">
+
+    <android.support.design.widget.FloatingActionButton
+        android:id="@+id/floating_action_button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="end|bottom"
+        android:layout_margin="16dp"
+        app:layout_constrainedHeight="true"
+        app:layout_constrainedWidth="true"
+        app:layout_editor_absoluteX="@dimen/activity_horizontal_margin"
+        app:layout_editor_absoluteY="@dimen/activity_horizontal_margin" />
+
+</android.support.constraint.ConstraintLayout>
