+++ research/widget/src/main/res/values/strings.xml
@@ -1,3 +1,5 @@
 <resources>
     <string name="app_name">widget</string>
+    <string name="title_tool_bar">Toolbar Title</string>
+    <string name="title_collapsing_tool_bar">我科姿势江山美如画</string>
 </resources>
